- fixed inconsistent data type used for protocol (int and unsigned short);   All fields carrying protocol or port moved unsigned short. More safer and   many issues related to casting eliminated.
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@1579 689a6050-402a-0410-94f2-e92a70836424

+++ msg_translator.c
@@ -147,7 +147,7 @@ extern int version_len;
  *  resolver = DO_DNS | DO_REV_DNS; if 0 no dns check is made
  * return 0 if equal */
 static int check_via_address(struct ip_addr* ip, str *name, 
-				unsigned short port, int proto, int resolver)
+				unsigned short port, unsigned short proto, int resolver)
 {
 	struct hostent* he;
 	int i;
